// src/shared/types/models/PointRead.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type PointRead = {
    id?: string;
    address: string;
    is_materials_delivered?: boolean;
    date_of_connection_of_the_point: string;
    date_of_issue_of_the_last_card: string;
    number_of_approved_requests?: number;
    number_of_cards_issued?: number;
    created_at?: string;
};